# stockrestaurante-openapi/src/main/resources/openapi-copilot.yaml
openapi: 3.0.3
info:
  title: Stock Restaurante API (Copilot)
  version: 1.0.0
  description: >
    Especificacion OpenAPI mantenida manualmente para el dominio actual de gestion
    de usuarios.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
tags:
  - name: Users
    description: Operaciones de gestion de usuarios.
paths:
  /api/users:
    get:
      tags: [Users]
      summary: Lista todos los usuarios
      operationId: listUsers
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Users]
      summary: Crea un usuario
      operationId: createUser
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpsertRequestDto'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: Identificador numerico del usuario.
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      tags: [Users]
      summary: Obtiene un usuario por id
      operationId: getUserById
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Users]
      summary: Actualiza un usuario existente
      operationId: updateUser
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpsertRequestDto'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Users]
      summary: Elimina un usuario
      operationId: deleteUser
      security:
        - basicAuth: []
      responses:
        '204':
          description: Usuario eliminado correctamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    BadRequest:
      description: Solicitud invalida por datos o formato incorrecto.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Credenciales ausentes o invalidas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no dispone de permisos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso solicitado no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto de negocio (estado incompatible o recurso duplicado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Error de validacion semantica en la solicitud.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error interno no controlado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UserDto:
      type: object
      required:
        - id
        - username
        - enabled
        - roles
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          example: 1
        username:
          type: string
          minLength: 1
          maxLength: 80
          example: admin
        enabled:
          type: boolean
          example: true
        roles:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
            maxLength: 80
            example: ROLE_ADMIN
    UserUpsertRequestDto:
      type: object
      required:
        - username
        - password
        - enabled
        - roles
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 80
          description: Nombre de usuario. Se normaliza internamente a minusculas.
          example: manager
        password:
          type: string
          minLength: 8
          maxLength: 120
          format: password
          example: securePass123
        enabled:
          type: boolean
          example: true
        roles:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
            maxLength: 80
            description: Rol del usuario; si no incluye prefijo ROLE_ se normaliza.
            example: ADMIN
    ValidationErrorItem:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: username
        message:
          type: string
          example: El nombre de usuario es obligatorio
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-09-25T08:20:58Z'
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: El id debe ser mayor que cero
        errors:
          type: array
          description: Detalle opcional de errores de validacion por campo.
          items:
            $ref: '#/components/schemas/ValidationErrorItem'